WIP: move commit.go into core.go
